+++ src/Core/Dfe.Complete.Application/Mappers/AutoMapping.cs
@@ -4,13 +4,16 @@ using Dfe.Complete.Domain.Entities;
 
 namespace Dfe.Complete.Application.Common.Mappers
 {
-    public sealed class AutoMapping : Profile
+	public sealed class AutoMapping : Profile
 	{
 		public AutoMapping()
 		{
 			CreateMap<Project, ProjectDto>();
 			CreateMap<ProjectGroup, ProjectGroupDto>();
 			CreateMap<User, UserDto>();
-        }
+			CreateMap<Note, NoteDto>()
+				.ForCtorParam(nameof(NoteDto.UserFullName),
+					opt => opt.MapFrom(src => src.User.FullName));
+		}
 	}
 }