+++ src/Core/Dfe.Complete.Application/Notes/Interfaces/INoteReadRepository.cs
@@ -1,0 +1,9 @@
+using Dfe.Complete.Application.Projects.Models;
+using Dfe.Complete.Domain.ValueObjects;
+
+namespace Dfe.Complete.Application.Notes.Interfaces;
+
+public interface INoteReadRepository
+{
+    IQueryable<NoteDto> GetNotesForProject(ProjectId projectId);
+}